+++ backend/src/main/java/com/carffeine/carffeine/auth/service/AuthService.java
@@ -3,8 +3,8 @@ package com.carffeine.carffeine.auth.service;
 import com.carffeine.carffeine.auth.domain.OAuthMember;
 import com.carffeine.carffeine.auth.domain.Provider;
 import com.carffeine.carffeine.auth.domain.TokenProvider;
-import com.carffeine.carffeine.member.domain.Member;
-import com.carffeine.carffeine.member.domain.MemberRepository;
+import com.carffeine.carffeine.member.domain.member.Member;
+import com.carffeine.carffeine.member.domain.member.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
