+++ backend/src/test/java/com/carffeine/carffeine/fake/member/FakeMemberRepository.java
@@ -1,7 +1,7 @@
 package com.carffeine.carffeine.fake.member;
 
-import com.carffeine.carffeine.member.domain.Member;
-import com.carffeine.carffeine.member.domain.MemberRepository;
+import com.carffeine.carffeine.member.domain.member.Member;
+import com.carffeine.carffeine.member.domain.member.MemberRepository;
 
 import java.util.HashMap;
 import java.util.Map;
